+++ src/stage1/gcc-libs-shim.sh
@@ -20,37 +20,30 @@
 
 set -eu
 
-_pkgname=glibc-shim
-_pkgver=$(pacman -Qi $_target-glibc | grep '^Version' | cut -d':' -f2 | tr -d [:space:])
+_pkgname=gcc-libs-shim
+_pkgver=$(pacman -Qi $_target-gcc | grep '^Version' | cut -d':' -f2 | tr -d [:space:])
 _pkgdir="$_makepkgdir"/$_pkgname/pkg/$_pkgname
 
 msg "makepkg: $_pkgname-$_pkgver-$_arch.pkg.tar.xz"
 
-mkdir -pv "$_makepkgdir"/$_pkgname
-pushd "$_makepkgdir"/$_pkgname >/dev/null
+if [ ! -f "$_makepkgdir"/$_pkgname-$_pkgver-$_arch.pkg.tar.xz ]; then
+  rm -rf "$_makepkgdir"/$_pkgname
+  mkdir -pv "$_makepkgdir"/$_pkgname
+  pushd "$_makepkgdir"/$_pkgname >/dev/null
 
-# to produce glibc shim from gcc, we need the package
-pacman -Sw --noconfirm --cachedir . $_target-glibc
-mkdir tmp && bsdtar -C tmp -xf $_target-glibc-$_pkgver-*.pkg.tar.xz
+  # to produce gcc-libs shim from gcc, we need the package
+  pacman -Sw --noconfirm --cachedir . $_target-gcc
+  mkdir tmp && bsdtar -C tmp -xf $_target-gcc-$_pkgver-*.pkg.tar.xz
 
-mkdir -p "$_pkgdir"/{etc,usr/{include,bin,lib,share}}
+  # install libraries
+  mkdir -p "$_pkgdir"/usr
+  cp -a tmp/usr/$_target/lib "$_pkgdir"/usr/lib
 
-# install binaries
-cp -a tmp/usr/$_target/{,usr/}bin/* "$_pkgdir"/usr/bin/
-# install libraries
-cp -a tmp/usr/$_target/{,usr/}lib/* "$_pkgdir"/usr/lib/
-# install headers
-cp -a tmp/usr/$_target/usr/include/* "$_pkgdir"/usr/include/
-rm -rf "$_pkgdir"/usr/include/scsi
-# install auxiliaries
-cp -a tmp/usr/$_target/etc/rpc "$_pkgdir"/etc/
-cp -a tmp/usr/$_target/usr/share/{i18n,locale} "$_pkgdir"/usr/share/
-
-# produce .PKGINFO file
-cat > "$_pkgdir"/.PKGINFO << EOF
+  # produce .PKGINFO file
+  cat > "$_pkgdir"/.PKGINFO << EOF
 pkgname = $_pkgname
 pkgver = $_pkgver
-pkgdesc = GNU C Library (extracted from $_target-glibc)
+pkgdesc = Runtime libraries shipped by GCC (extracted from $_target-gcc)
 url = https://github.com/riscv/riscv-gnu-toolchain
 builddate = $(date '+%s')
 size = $(( $(du -sk --apparent-size "$_pkgdir" | cut -d'	' -f1) * 1024 ))
@@ -59,18 +52,21 @@ provides = ${_pkgname%-*}
 conflicts = ${_pkgname%-*}
 EOF
 
-# package
-cd "$_pkgdir"
-env LANG=C bsdtar -czf .MTREE \
-  --format=mtree \
-  --options='!all,use-set,type,uid,gid,mode,time,size,md5,sha256,link' \
-  .PKGINFO *
-env LANG=C bsdtar -cf - .MTREE .PKGINFO * | xz -c -z - > \
-  "$_chrootdir"/packages/$_arch/$_pkgname-$_pkgver-$_arch.pkg.tar.xz
+  # package
+  cd "$_pkgdir"
+  env LANG=C bsdtar -czf .MTREE \
+    --format=mtree \
+    --options='!all,use-set,type,uid,gid,mode,time,size,md5,sha256,link' \
+    .PKGINFO *
+  env LANG=C bsdtar -cf - .MTREE .PKGINFO * | xz -c -z - > \
+    "$_makepkgdir"/$_pkgname-$_pkgver-$_arch.pkg.tar.xz
 
-popd >/dev/null
+  popd >/dev/null
 
-# rm -rf "$_makepkgdir"/$_pkgname
+  # rm -rf "$_makepkgdir"/$_pkgname
+fi
+
+cp -av "$_makepkgdir"/$_pkgname-$_pkgver-$_arch.pkg.tar.xz "$_chrootdir"/packages/$_arch
 
 rm -rf "$_chrootdir"/var/cache/pacman/pkg/*
 rm -rf "$_chrootdir"/packages/$_arch/repo.{db,files}*
