+++ src/stage1/glibc-shim.sh
@@ -20,29 +20,40 @@
 
 set -eu
 
-_pkgname=ca-certificates-utils-shim
-_pkgver=$(pacman -Qi ${_pkgname%-*} | grep '^Version' | cut -d':' -f2 | tr -d [:space:])
+_pkgname=glibc-shim
+_pkgver=$(pacman -Qi $_target-glibc | grep '^Version' | cut -d':' -f2 | tr -d [:space:])
 _pkgdir="$_makepkgdir"/$_pkgname/pkg/$_pkgname
 
 msg "makepkg: $_pkgname-$_pkgver-$_arch.pkg.tar.xz"
 
-mkdir -pv "$_makepkgdir"/$_pkgname
-pushd "$_makepkgdir"/$_pkgname >/dev/null
+if [ ! -f "$_makepkgdir"/$_pkgname-$_pkgver-$_arch.pkg.tar.xz ]; then
+  rm -rf "$_makepkgdir"/$_pkgname
+  mkdir -pv "$_makepkgdir"/$_pkgname
+  pushd "$_makepkgdir"/$_pkgname >/dev/null
 
-# get original package
-pacman -Sw --noconfirm --cachedir . ${_pkgname%-*}
-mkdir tmp && bsdtar -C tmp -xf ${_pkgname%-*}-$_pkgver-*.pkg.tar.xz
+  # to produce glibc shim from gcc, we need the package
+  pacman -Sw --noconfirm --cachedir . $_target-glibc
+  mkdir tmp && bsdtar -C tmp -xf $_target-glibc-$_pkgver-*.pkg.tar.xz
 
-# install certificates
-mkdir -p "$_pkgdir"/etc/ssl/certs/
-cp -a tmp/etc/ssl/certs/ca-certificates.crt "$_pkgdir"/etc/ssl/certs/.
+  mkdir -p "$_pkgdir"/{etc,usr/{include,bin,lib,share}}
 
-# produce .PKGINFO file
-cat > "$_pkgdir"/.PKGINFO << EOF
+  # install binaries
+  cp -a tmp/usr/$_target/{,usr/}bin/* "$_pkgdir"/usr/bin/
+  # install libraries
+  cp -a tmp/usr/$_target/{,usr/}lib/* "$_pkgdir"/usr/lib/
+  # install headers
+  cp -a tmp/usr/$_target/usr/include/* "$_pkgdir"/usr/include/
+  rm -rf "$_pkgdir"/usr/include/scsi
+  # install auxiliaries
+  cp -a tmp/usr/$_target/etc/rpc "$_pkgdir"/etc/
+  cp -a tmp/usr/$_target/usr/share/{i18n,locale} "$_pkgdir"/usr/share/
+
+  # produce .PKGINFO file
+  cat > "$_pkgdir"/.PKGINFO << EOF
 pkgname = $_pkgname
 pkgver = $_pkgver
-pkgdesc = Common CA certificates (utilities) - extracted from host machine
-url = http://pkgs.fedoraproject.org/cgit/rpms/ca-certificates.git
+pkgdesc = GNU C Library (extracted from $_target-glibc)
+url = https://github.com/riscv/riscv-gnu-toolchain
 builddate = $(date '+%s')
 size = $(( $(du -sk --apparent-size "$_pkgdir" | cut -d'	' -f1) * 1024 ))
 arch = $_arch
@@ -50,21 +61,23 @@ provides = ${_pkgname%-*}
 conflicts = ${_pkgname%-*}
 EOF
 
-# package
-cd "$_pkgdir"
-env LANG=C bsdtar -czf .MTREE \
-  --format=mtree \
-  --options='!all,use-set,type,uid,gid,mode,time,size,md5,sha256,link' \
-  .PKGINFO *
-env LANG=C bsdtar -cf - .MTREE .PKGINFO * | xz -c -z - > \
-  "$_chrootdir"/packages/$_arch/$_pkgname-$_pkgver-$_arch.pkg.tar.xz
+  # package
+  cd "$_pkgdir"
+  env LANG=C bsdtar -czf .MTREE \
+    --format=mtree \
+    --options='!all,use-set,type,uid,gid,mode,time,size,md5,sha256,link' \
+    .PKGINFO *
+  env LANG=C bsdtar -cf - .MTREE .PKGINFO * | xz -c -z - > \
+    "$_makepkgdir"/$_pkgname-$_pkgver-$_arch.pkg.tar.xz
 
-popd >/dev/null
+  popd >/dev/null
 
-# rm -rf "$_makepkgdir"/$_pkgname
+  # rm -rf "$_makepkgdir"/$_pkgname
+fi
+
+cp -av "$_makepkgdir"/$_pkgname-$_pkgver-$_arch.pkg.tar.xz "$_chrootdir"/packages/$_arch
 
 rm -rf "$_chrootdir"/var/cache/pacman/pkg/*
 rm -rf "$_chrootdir"/packages/$_arch/repo.{db,files}*
 repo-add -R "$_chrootdir"/packages/$_arch/{repo.db.tar.gz,*.pkg.tar.xz}
 pacman --noconfirm --config "$_chrootdir"/etc/pacman.conf -r "$_chrootdir" -Syy $_pkgname
-
